Fixed memory leak in CrossSocket.
Before, TAbstractCrossSocket never destroyed because there is a circular reference with the class TIoEventThread, so I added the Weak. The change in AfterConstruction is because inside of it, after Inherited, RefCount decreases to 0, and calling StartLoop or any method after Inherited within AfterConstruction will be extremely dangerous, since any reference that increases within these methods, when decreasing and internally checking that the ReCount is 0, it will destroy the TAbstractCrossSocket before it even completely exits the AfterConstruction.