Update to the Delphi 10.4 (Sydney)
In Delphi 10.4, have a compilation error in Android32 because the socklen_t is Integer in Android32 and Cardinal in others platforms; I don't know if was just this issue with new delphi, because I'm still doing the tests with CrossSocket in new delphi 10.4 Sydney. But compilations erros is just this.